+++ hw06-spring_ORM_JPA/src/main/java/ru/otus/services/BookService.java
@@ -15,7 +15,7 @@ public interface BookService {
     void update(long bookId, String bookName) throws Exception;
     List<Book> findAll();
     BookDto findById(long id) throws Exception;
-    int getCountBooks();
+    Long getCountBooks();
     List<Book> getBooksByAuthor(long authorId) throws Exception;
     List<Book> getBooksByGenre(long genreId) throws Exception;
 }
